refactored helpers file
